Change the type of `positions` and add `positions()`
`MatchedPath::positions()` will be used to emphasize what characters
matches in a relative path. In order to expose this function, I changed
the type of the field `positions` from `VecDeque` to `Vec`.

`VecDeque` was still used to calculates positions, but the result will
be stored as `Vec` from this patch.