Change the type of `positions` and add `positions()` (#47)
MatchedPath::positions() will be used to emphasize what characters
matches in a relative path. In order to expose this function, I changed
the type of the field positions from VecDeque to Vec. I wanted
to use as_slice instead of as_slices, and I'd like to avoid calling
make_contiguous.

VecDeque was still used to calculates positions, but the result will
be stored as Vec from this patch.