Add a new struct: StartingPoint (#68)
This struct is intended for a caller of `Finder` to specify the correct
`starting_point` in advance. **The correct `starting_point`** means the
path which is valid in terms of character encdoing, and really exists in
a file system; in other words, it can be `canonicalize`d safely.

With this refactoring, each `ConsumedDir` does not have to call
`canonicalize` because `Finder` can get absolute path due to the
`canonicalize`d `starting_point`.