Returns Chunks based on the terminal size
Sometimes, a user may come across the case that some of the file
candidates are too long for the terminal to wrap them and cause the
broken screen. This patch tries to alleviate such a situation to omit
them with an ellipsis. Some of the matched characters will be lost
because of the omission, but I want to prioritize to maintain the screen
not broken.