Add --status-line=<absolute|relative|none>
Users might want to see the relative path on the status instead of the
absolute path. This option allows them to show the relative path.
In addition, I added another option value "none" to hide the status-line
itself.

Honestly, this option is intended for testing use, but I hope this would
be convenient.