Return usize of cjk width when query position is changed
When Query's value is changed, a terminal usually has to move cursor.
At this time, the terminal should know how many cells should it move.
If the last value was emoji, the terminal should move the cursor by 2
most of time.

This patch makes `Query` return the number of cells should be moved.