Add Screen to replace cli::Runner (#1040)
`cli::Runner` is a bit difficult to update for adding new features,
so I want to replace it with the new implementation, `Screen`.

`Screen` will behave like `cli::Runner`, but it's written for
extensibility. Still, this struct is huge, but I believe it's better
than `cli::Runner`.

This patch just adds `Screen`. I will replace `cli::Runner` with it
later.