Remove the tests/ directory
Previously, test cases in the `tests/` directory were difficult to
update because they compare the output bytes one by one. If we add a new
feture and change the output, the tests will fail. This is cumbersome to
change the existing test cases.

I think smoke tests should be done manually, not by unit testing, so I
decided to remove the `tests/` directory.