Add Tree to store all tree files in memory (#1059)
The current implementation has a performance problem because it always
get all files when the passed `query` changes. This significantly
increases I/O. I want to avoid the call of I/O by storing all files in
memory. This development decision might affect the memory usage,
but I want to first try it.

This patch just adds `Tree` for the purpose of storing all possible
files, which is not yet used by other modules.