Support Ctrl-R to reload the file tree
Although the current thwack implementation loads file tree in memory at
the beginning of this program, sometimes a user might want to reload the
file tree for some reason. I think it's worth adding a new feature to
reload the file tree when the user commands to reload it.

This patch allows a user to reload the file tree when the user type
Ctrl-R.